Refactor breaker structure, fix circular imports
