+++ conftest.py
@@ -46,7 +46,7 @@ pytest_plugins = [
     "saleor.account.tests.fixtures",
     "saleor.graphql.account.tests.fixtures",
     "saleor.payment.tests.fixtures",
-    "saleor.webhook.transport.tests.circuit_breaker.fixtures",
+    "saleor.webhook.tests.circuit_breaker.fixtures",
 ]
