+++ saleor/graphql/app/types.py
@@ -70,7 +70,7 @@ from .resolvers import (
 # TODO: Remove the conditional when unit tests circular import is solved.
 breaker_board = None
 if settings.ENABLE_BREAKER_BOARD:
-    from ...webhook.transport.synchronous.circuit_breaker.breaker_board import (
+    from ...webhook.circuit_breaker.breaker_board import (
         initialize_breaker_board,
     )
